FFA Code Standardization
